pyproject.toml
[tool.poetry]
name = "django-project-template"
version = "0.1.0"
description = "My django template"
authors = ["Lucho Leves <lucianoleves@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6.1"
django = "^2.1"
django-configurations = "^2.2"

[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.2.1"
pytest = "^6.2.3"
factory-boy = "^3.2.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
